ecr_discord_bot/logic_cloud_function_python/requirements.txt
cryptography==3.3.1
pyjwt==2.7.0
boto3==1.26.149
google-cloud-compute==1.16.1